Move get_bucket_name into helper module
