+++ dmscripts/bulk_upload_documents.py
@@ -4,13 +4,6 @@ import csv
 
 from dmutils.documents import get_document_path, generate_download_filename
 
-BUCKET_CATEGORIES = [
-    'agreements',
-    'communications',
-    'documents',
-    'submissions'
-]
-
 
 def upload_file(bucket, dry_run, file_path, framework_slug, bucket_category, supplier_name_dict=None):
     # Retrieve the supplier ID from the filepath
@@ -43,19 +36,6 @@ def upload_file(bucket, dry_run, file_path, framework_slug, bucket_category, sup
         print(supplier_id)
     else:
         print("[Dry-run] UPLOAD: '{}' to '{}'".format(file_path, upload_path))
-
-
-def get_bucket_name(stage, bucket_category):
-    if bucket_category not in BUCKET_CATEGORIES:
-        return None
-    if stage in ['local', 'dev']:
-        return "digitalmarketplace-dev-uploads"
-    if stage not in ['preview', 'staging', 'production']:
-        return None
-
-    bucket_name = 'digitalmarketplace-{0}-{1}-{1}'.format(bucket_category, stage)
-    print("BUCKET: {}".format(bucket_name))
-    return bucket_name
 
 
 def get_all_files_of_type(local_directory, file_type):
