+++ tests/test_bulk_upload_documents.py
@@ -7,33 +7,11 @@ try:
 except ImportError:
     import builtins
 
-from dmscripts.bulk_upload_documents import get_bucket_name, get_all_files_of_type, \
+from dmscripts.bulk_upload_documents import get_all_files_of_type, \
     get_supplier_id_from_framework_file_path, upload_file, get_document_name_from_file_path, \
     get_supplier_name_dict_from_tsv
 import mock
 import pytest
-
-
-class TestGetBucketName:
-
-    @pytest.mark.parametrize(
-        'stage, expected_bucket_name',
-        [
-            ('local', 'digitalmarketplace-dev-uploads'),
-            ('dev', 'digitalmarketplace-dev-uploads'),
-            ('preview', 'digitalmarketplace-agreements-preview-preview'),
-            ('staging', 'digitalmarketplace-agreements-staging-staging'),
-            ('production', 'digitalmarketplace-agreements-production-production'),
-        ]
-    )
-    def test_get_bucket_name_for_agreements_documents(self, stage, expected_bucket_name):
-        assert get_bucket_name(stage, 'agreements') == expected_bucket_name
-
-    def test_get_bucket_name_returns_none_for_invalid_stage(self):
-        assert get_bucket_name('xanadu', 'agreements') is None
-
-    def test_get_bucket_name_returns_none_for_invalid_bucket_category(self):
-        assert get_bucket_name('local', 'bananas') is None
 
 
 class TestGetAllFilesOfType:
